+++ packages/ng-primitives/menu/src/menu/menu.ts
@@ -25,6 +25,8 @@ import { NgpMenuToken, provideMenu } from './menu-token';
     '[style.top.px]': 'overlay.position().y',
     '[style.--ngp-menu-trigger-width.px]': 'overlay.triggerWidth()',
     '[style.--ngp-menu-transform-origin]': 'overlay.transformOrigin()',
+    '[attr.data-placement]': 'overlay.finalPlacement()',
+    'data-overlay': '',
   },
 })
 export class NgpMenu {